% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ils_}
\alias{ils_}
\title{ils
Lists collections and data objects in iRODS}
\usage{
ils_(args)
}
\arguments{
\item{args}{"ils" one or several iRODS data-object (file) or collection (directory) paths}
}
\value{
A data frame with information about collections and data objects under the supplied paths
}
\description{
ils
Lists collections and data objects in iRODS
}
