\name{rirods-package}
\alias{rirods-package}
\alias{rirods}
\docType{package}
\title{
\packageTitle{rirods}
}
\description{
\packageDescription{rirods}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rirods}
\packageIndices{rirods}
%%~~ An overview of how to use the package,
%%   including the most important functions ~~
}
\author{
\packageAuthor{rirods}

Maintainer: \packageMaintainer{rirods}
}

%%\references{
%%~~ Literature or other references for background information ~~
%%}
%%~~ Optionally other standard keywords, one per line,
%%   from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
